requirements.txt
ccxt>=4.0.0
pandas>=1.5.0
numpy>=1.23.0
statsmodels>=0.14.0
requests>=2.28.0
python-dotenv>=1.0.0

# Dev & testing
pytest>=7.0.0
pytest-cov>=4.0.0
